Moving CSRFToken and fetchData to utils
